moved ALL_PARAMS_KEY_VALUE_REGEX to props file
