+++ wiggle_query_language/clauses/regexes/make/make_patterns.py
@@ -38,14 +38,14 @@ MAKE_STATEMENT_CHECK_PARAMS_SYNTAX_REGEX = compile(
 
 
 # int: 1, float: 3.14, bool: true, none: null, str: '2', str2:"2_4", str3: "3 4 5", email: 'foo@bar.net',  list: [...]
-ALL_PARAMS_KEY_VALUE_REGEX = compile(
-    (
-        r"(?P<property_name>\w+)\s*:\s*(?P<property_value>(?P<none_type>null)|(?P<bool_type>true|false)|"
-        r"(?P<float_type>\d+\.\d+)|(?P<int_type>[0-9]+)|(?P<list_type>\[[\w,\s'\"\.@\+]+\])|"
-        r"(?P<string_type>[\w+\'\"@\.\s]+))"
-    ),
-    flags=IGNORECASE,
-)
+# ALL_PARAMS_KEY_VALUE_REGEX = compile(
+#     (
+#         r"(?P<property_name>\w+)\s*:\s*(?P<property_value>(?P<none_type>null)|(?P<bool_type>true|false)|"
+#         r"(?P<float_type>\d+\.\d+)|(?P<int_type>[0-9]+)|(?P<list_type>\[[\w,\s'\"\.@\+]+\])|"
+#         r"(?P<string_type>[\w+\'\"@\.\s]+))"
+#     ),
+#     flags=IGNORECASE,
+# )
 
 
 # [1, '2', "2_4", "3 4", 3.14]
