+++ wiggle_query_language/clauses/regexes/make/tests/test_make_re.py
@@ -4,7 +4,6 @@ import pytest
 
 from testing.test_helpers import does_not_raise
 from wiggle_query_language.clauses.regexes.make.make_patterns import (
-    ALL_PARAMS_KEY_VALUE_REGEX,
     ILLEGAL_CHARS_REGEX,
     MAKE_STATEMENT_ALL_REGEX,
     MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX,
@@ -120,54 +119,6 @@ class TestMakeRePatterns:
             test = MAKE_STATEMENT_CHECK_PARAMS_SYNTAX_REGEX.findall(test_pattern)
             assert test == expected_result
 
-    # @pytest.mark.parametrize(
-    #     "test_pattern, expected_result, exception",
-    #     [
-    #         pytest.param(
-    #             """MAKE (:NodeLabel)-[f:REL]->(foo:NodeLabel);""",
-    #             [("-[f:REL]->", "REL")],
-    #             does_not_raise(),
-    #             id="EXP PASS: 1 Match",
-    #         ),
-    #         pytest.param(
-    #             """MAKE (:NodeLabel)---[f:REL]--->(foo:NodeLabel);""",
-    #             [("---[f:REL]--->", "REL")],
-    #             does_not_raise(),
-    #             id="EXP PASS: 1 Match, long rel",
-    #         ),
-    #         pytest.param(
-    #             """MAKE (:NodeLabel)-[]-(foo:NodeLabel);""",
-    #             [("-[]-", "")],
-    #             does_not_raise(),
-    #             id="EXP PASS: 1 Match",
-    #         ),
-    #         pytest.param(
-    #             """MAKE (:NodeLabel)----[]---(foo:NodeLabel);""",
-    #             [("----[]---", "")],
-    #             does_not_raise(),
-    #             id="EXP PASS: 1 Match, long rel not symmetrical",
-    #         ),
-    #         pytest.param(
-    #             """MAKE (:NodeLabel)<-[f:FOO]->(foo:NodeLabel);""",
-    #             [("<-[f:FOO]->", "FOO")],
-    #             does_not_raise(),
-    #             id="EXP PASS: 1 Match",
-    #         ),
-    #         pytest.param(
-    #             "REPORT (:NodeLabel);",
-    #             [],
-    #             does_not_raise(),
-    #             id="EXP PASS: No Match",
-    #         ),
-    #     ],
-    # )
-    # def test_relationship_dir_check_regex(
-    #     self, test_pattern: str, expected_result: Optional[list[str]], exception
-    # ) -> None:
-    #     with exception:
-    #         test = RELATIONSHIP_DIR_CHECK_REGEX.findall(test_pattern)
-    #         assert test == expected_result
-
     @pytest.mark.parametrize(
         "test_pattern, expected_result, exception",
         [
@@ -214,49 +165,4 @@ class TestMakeRePatterns:
     ) -> None:
         with exception:
             test = ILLEGAL_CHARS_REGEX.findall(test_pattern)
-            assert test == expected_result
-
-    @pytest.mark.parametrize(
-        "test_pattern, expected_result, exception",
-        [
-            pytest.param(
-                """{int: 1, float: 3.14, bool: true, bool2: false, none: null, str: '2', str2:"2_4", str3: "3 4 5", email: 'foo@bar.net', pn: '+447541254566', list: [1, 3.14, true, false, '2', "2_4", "3 4", "foo@bar.net", "+447541254566"]}""",
-                [
-                    ("int", "1", "", "", "", "1", "", ""),
-                    ("float", "3.14", "", "", "3.14", "", "", ""),
-                    ("bool", "true", "", "true", "", "", "", ""),
-                    ("bool2", "false", "", "false", "", "", "", ""),
-                    ("none", "null", "null", "", "", "", "", ""),
-                    ("str", "'2'", "", "", "", "", "", "'2'"),
-                    ("str2", '"2_4"', "", "", "", "", "", '"2_4"'),
-                    ("str3", '"3 4 5"', "", "", "", "", "", '"3 4 5"'),
-                    ("email", "'foo@bar.net'", "", "", "", "", "", "'foo@bar.net'"),
-                    ("pn", "'+447541254566'", "", "", "", "", "", "'+447541254566'"),
-                    (
-                        "list",
-                        '[1, 3.14, true, false, \'2\', "2_4", "3 4", "foo@bar.net", "+447541254566"]',
-                        "",
-                        "",
-                        "",
-                        "",
-                        '[1, 3.14, true, false, \'2\', "2_4", "3 4", "foo@bar.net", "+447541254566"]',
-                        "",
-                    ),
-                ],
-                does_not_raise(),
-                id="EXP PASS: No Match",
-            ),
-            pytest.param(
-                "{int: 1}",
-                [("int", "1", "", "", "", "1", "", "")],
-                does_not_raise(),
-                id="EXP PASS: No Match",
-            ),
-        ],
-    )
-    def test_all_params_key_value_regex(
-        self, test_pattern: str, expected_result: Optional[list[str]], exception
-    ) -> None:
-        with exception:
-            test = ALL_PARAMS_KEY_VALUE_REGEX.findall(test_pattern)
             assert test == expected_result
